+++ pub/.rubocop.yml
@@ -1,1 +1,4 @@
 inherit_from: ../.rubocop.yml
+
+Sorbet/TrueSigil:
+  Enabled: true
